LMesh.jl

*Basic building block for other codes *

Basic types

LMesh.Mesh — Type

Supertype for Mesh2D and Mesh3D

LMesh.Mesh2D — Type

Basic structure for 2D Meshes

Mesh2D(bmesh::Bmesh2D,materials::Vector{Material}, geometries::Vector{Geometry},ebc::Matrix{Float64}, nbc::Matrix{Float64} ; matele = Int64[], geoele = Int64[], options=Dict{Symbol,Matrix{Float64}}())

where <br/>

bmesh = Background mesh (2D) <br/>

materials = Vector with Materials definitions

geometries = Vector with Geometries definitions

ebc = Matrix with essential boundary conditions

nbc = Matrix with natural boundary conditions

mat_ele = Vector linking each element to a given material

geo_ele = Vector linking each element to a given geometry

options = Dictionary with optional data

LMesh.Mesh3D — Type

Basic structure for #D Meshes

Mesh3D(bmesh::Bmesh3D,materials::Vector{Material}, geometries::Vector{Geometry},ebc::Matrix{Float64}, nbc::Matrix{Float64} ; matele = Int64[], geoele = Int64[], options=Dict{Symbol,Matrix{Float64}}())

where

bmesh = Background mesh (3D)

materials = Vector with Materials definitions

geometries = Vector with Geometries definitions

ebc = Matrix with essential boundary conditions

nbc = Matrix with natural boundary conditions

mat_ele = Vector linking each element to a given material

geo_ele = Vector linking each element to a given geometry

options = Dictionary with optional data

Iterator

Base.iterate — Function

Iterator for Mesh. Iterates from 1 to the number of elements in the mesh. To be used in loops like for ele in m ... end

Material

LMesh.Material — Type

Basic structure to store material data

Material(;Ex=1.0,νxy=0.0, α=0.0, density=1.0, model=:EPT, custom=[1.0 1.0 ; 1.0 1.0])

where all inputs are named.

Geometry

LMesh.Geometry — Type

Basic structure to store geometrical data

Geometry(;A=1.0,thickness=1.0)

where all inputs are named.

Base

LMesh.Free_DOFs — Function

Return a vector with all the free (not constrained) DOFs in the mesh. This function is used to build Mesh.free_dofs

Free_DOFs(bmesh::Bmesh,nebc::Int64,ebc::Matrix{Float64})

where

bmesh = background mesh nebc = number of essential boundary conditions ebc = matrix with essential boundary condition data

LMesh.Nodal_coordinates — Function

Return vectors x and y with the nodal coordinates of element ele

Nodal_coordinates(m::Mesh2D,ele)

where m is a 2D Mesh and ele a valid element.

Return vectors x, y and z with the nodal coordinates of element ele

Nodal_coordinates(m::Mesh3D,ele)

where m is a 3D Mesh and ele a valid element.

LMesh.Centroid — Function

Return a vector with the centroidal coordinates of element ele

Centroid(mesh::Mesh2D,ele::Int64)

Return a vector with the centroidal coordinates of element ele

Centroid(mesh::Mesh3D,ele::Int64)

LMesh.Get_dim — Function

Return 2 for 2D problems and 3 for 3D problems

Get_dim(mesh::Mesh)

LMesh.Get_etype — Function

Return element type

Get_etype(mesh::Mesh)

LMesh.Get_eclass — Function

Return truss or solid

Get_eclass(mesh::Mesh)

LMesh.Get_material_number — Function

Material number for element ele

Get_material(mesh::Mesh,ele::Int64)

LMesh.Get_material — Function

Material data for element ele

Get_material(mesh::Mesh,ele::Int64)

LMesh.Get_geometry_number — Function

Geometry number for element ele

Getgeometrynumber(mesh::Mesh,ele::Int64)

LMesh.Get_geometry — Function

Geometry data for element ele

Get_geometry(mesh::Mesh,ele::Int64)

LMesh.Get_ne — Function

Number of elements in this mesh

Get_ne(mesh::Mesh)

LMesh.Get_nn — Function

Number of nodes in this mesh

Get_nn(mesh::Mesh)

Overloaded from BMesh

BMesh.Connect — Function

(Overloaded from BMesh) Return the connectivities of element ele

Connect(mesh::Mesh,ele::Int64)

as an Int64 vector.

BMesh.Coord — Function

(Overloaded from BMesh) Return the coordinates of node

Coord(mesh::Mesh,node::Int64)

as a vector [x;y;(z)]

BMesh.Length — Function

(Overloaded from BMesh) Return the distance between two nodes of the element

Length(mesh::Mesh,ele::Int64;nodes=(1,2))

where the default is the distance between (local) nodes 1 and 2

BMesh.DOFs — Function

(Overloaded from BMesh) Return a dim*n vector with the global DOFs of element ele

DOFs(mesh::Mesh,ele::Int64)

where n is the number of nodes of ele and dim=2 or 3.

BMesh.T_matrix — Function

(Overloaded from BMesh) Evaluate the rotation matrix T that maps from global to local reference systems. This version evaluates Rotation internally.

T_matrix(mesh::Mesh, ele::Int64, α=0.0)

Plot

RecipesBase.plot — Function

Overload plot to show the mesh and boundary conditions. Just for truss elements.

plot(mesh::Mesh2D; name="", factor=6.0)

factor is used to scale arrows to a fraction of the total dimension.